+++ packages/app-config-writer/test/unit/cards-config/cap.test.ts
@@ -1,16 +1,50 @@
 import { jest } from '@jest/globals';
-import { join } from 'node:path';
+import { join, dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
 import { create as createStorage } from 'mem-fs';
 import { create } from 'mem-fs-editor';
 
+const mockGetProjectType = jest.fn();
+const mockFindCapProjectRoot = jest.fn();
+const mockReadManifest = jest.fn();
+
+const realProjectAccess = await import('@sap-ux/project-access');
+
+jest.unstable_mockModule('@sap-ux/project-access', () => ({
+    ...realProjectAccess,
+    getProjectType: mockGetProjectType,
+    findCapProjectRoot: mockFindCapProjectRoot
+}));
+
+const realUtils = await import('../../../src/common/utils.js');
+
+jest.unstable_mockModule('../../../src/common/utils.js', () => ({
+    ...realUtils,
+    readManifest: mockReadManifest
+}));
+
+jest.unstable_mockModule('../../../src/cards-config/prerequisites.js', () => ({
+    ensureMinUI5Version: jest.fn().mockResolvedValue(undefined)
+}));
+
+const realUi5Yaml = await import('../../../src/common/ui5-yaml.js');
+
+jest.unstable_mockModule('../../../src/common/ui5-yaml.js', () => ({
+    ...realUi5Yaml,
+    updateMiddlewaresForPreview: jest.fn().mockResolvedValue(undefined)
+}));
+
+const { enableCardGeneratorConfig } = await import('../../../src/cards-config/index.js');
 const { updateCapRootPackageJsonForCards } = await import('../../../src/cards-config/cap.js');
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 const CAP_ROOT = '/cap-root';
 const APP_PATH = join(CAP_ROOT, 'app', 'my_app');
 const APP_ID = 'ns.myapp';
 const APP_FOLDER_NAME = 'my_app';
 
-function createTestFs(options?: { cardGeneratorPath?: string }) {
+function createCapTestFs(options?: { cardGeneratorPath?: string }) {
     const fs = create(createStorage());
     const cardGeneratorPath = options?.cardGeneratorPath ?? '/test/flpCardGeneratorSandbox.html';
     fs.write(
@@ -35,7 +69,7 @@ server:
 
 describe('updateCapRootPackageJsonForCards', () => {
     test('writes cds watch script to CAP root package.json with default path and intent', async () => {
-        const fs = createTestFs();
+        const fs = createCapTestFs();
 
         await updateCapRootPackageJsonForCards(CAP_ROOT, APP_ID, APP_FOLDER_NAME, APP_PATH, fs);
 
@@ -46,7 +80,7 @@ describe('updateCapRootPackageJsonForCards', () => {
     });
 
     test('writes cds watch script with custom card generator path from yaml', async () => {
-        const fs = createTestFs({ cardGeneratorPath: '/test/myCustomCardGen.html' });
+        const fs = createCapTestFs({ cardGeneratorPath: '/test/myCustomCardGen.html' });
 
         await updateCapRootPackageJsonForCards(CAP_ROOT, APP_ID, APP_FOLDER_NAME, APP_PATH, fs);
 
@@ -57,7 +91,7 @@ describe('updateCapRootPackageJsonForCards', () => {
     });
 
     test('uses appFolderName as script key suffix', async () => {
-        const fs = createTestFs();
+        const fs = createCapTestFs();
 
         await updateCapRootPackageJsonForCards(CAP_ROOT, APP_ID, 'another_app', APP_PATH, fs);
 
@@ -66,7 +100,7 @@ describe('updateCapRootPackageJsonForCards', () => {
     });
 
     test('preserves existing scripts in CAP root package.json', async () => {
-        const fs = createTestFs();
+        const fs = createCapTestFs();
         fs.writeJSON(join(CAP_ROOT, 'package.json'), {
             name: 'cap-project',
             scripts: { 'watch-my_app': 'cds watch --open my_app/webapp/index.html' }
@@ -86,5 +120,48 @@ describe('updateCapRootPackageJsonForCards', () => {
         await expect(
             updateCapRootPackageJsonForCards(CAP_ROOT, APP_ID, APP_FOLDER_NAME, APP_PATH, fs)
         ).rejects.toThrow(`package.json not found at CAP root: ${CAP_ROOT}`);
+    });
+});
+
+describe('enableCardGeneratorConfig - CAP routing', () => {
+    const basePath = join(__dirname, '../../fixtures/cards-config/lrop-v4');
+    const yamlPath = join(basePath, 'ui5.yaml');
+    const capRoot = '/test-cap-root';
+
+    beforeEach(() => {
+        jest.resetAllMocks();
+        mockReadManifest.mockResolvedValue({ manifest: { 'sap.app': { id: 'apps.v4.example' } } });
+        mockFindCapProjectRoot.mockResolvedValue(capRoot);
+    });
+
+    test('throws for CAPJava projects', async () => {
+        mockGetProjectType.mockResolvedValue('CAPJava');
+
+        await expect(enableCardGeneratorConfig(basePath, yamlPath)).rejects.toThrow(
+            'The cards-editor command is not supported for CAP Java projects.'
+        );
+    });
+
+    test('writes cds watch script to cap root for CAPNodejs', async () => {
+        mockGetProjectType.mockResolvedValue('CAPNodejs');
+        const fs = create(createStorage());
+        fs.writeJSON(join(capRoot, 'package.json'), { name: 'cap-project' });
+
+        await enableCardGeneratorConfig(basePath, yamlPath, undefined, fs);
+
+        const pkg = fs.readJSON(join(capRoot, 'package.json')) as Record<string, unknown>;
+        const scripts = pkg.scripts as Record<string, string>;
+        expect(scripts['start-cards-generator-lrop-v4']).toMatch(/^cds watch --open "/);
+    });
+
+    test('does not write to cap root for EDMXBackend', async () => {
+        mockGetProjectType.mockResolvedValue('EDMXBackend');
+        const fs = create(createStorage());
+        fs.writeJSON(join(capRoot, 'package.json'), { name: 'cap-project' });
+
+        await enableCardGeneratorConfig(basePath, yamlPath, undefined, fs);
+
+        const pkg = fs.readJSON(join(capRoot, 'package.json')) as Record<string, unknown>;
+        expect(pkg.scripts).toBeUndefined();
     });
 });
